d06_人脸检测与识别/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-multipart==0.0.9
facenet-pytorch==2.5.3
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4
pillow==10.3.0
pandas==2.2.1
